Separate logic into new file
